Create CodeBreakingGame
This class makes a code-breaking game. It asks one player to input a 4 digit number. Then, it stores each of these digits in separate cells of a matrix of length 4. The program outputs enough blank lines so that player 2 cannot see the 4 digit number. Until the correct number is input, the program takes a guess from player 2 and determines how many of the 4 digits are correct. The number of correct digits will be reported
to player 2 and they will be asked for the next guess. When the player finally solves the puzzle, the program outputs the number of tries needed.