Create CodeBreakingGame
This class makes a code-breaking game. The program asks one player to input a 4 digit number. Then, each of these digits is stored in a separate cell of a matrix of length 4. The program outputs enough blank lines so that player 2 cannot see the 4 digit number. Until the correct number is input, the program takes a guess from player 2 and determines how many of the four digits are correct. The number of correct digits is reported
to player 2 and they are asked for the next guess. When the player finally solves the puzzle, the program outputs the number of tries needed.